+++ CHANGELOG.md
@@ -23,7 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - org.springframework.security:spring-security-web to 6.1.1
 - Changed Dependencies file with new library versions.
 - Changed Mapping to adapt new Data Model on BPDM Gate
-- Enable Hidden endpoints for Sharing Bpn data
+- Enable Hidden endpoints for Sharing Controller
 
 
 ### Fixes